+++ app/Http/Controllers/transaksi_penjualancontroller.php
@@ -41,9 +41,11 @@ class transaksi_penjualancontroller extends Controller
         //     'detail' => $detail,
         //     // 'grand_total' => $grandTotal,
         // ]);
+
+       // return view("transaksi.penjualan.sales");
     }
 
-    public function createTrans($transaksiId)
+    public function create($transaksiId)
     {
         if (is_null($transaksiId)) {
             abort(404);
@@ -63,7 +65,7 @@ class transaksi_penjualancontroller extends Controller
         ]);
     }
 
-    public function storeTrans(TranSaleRequest $request)
+    public function store(TranSaleRequest $request)
     {
         $request = $request->all();
 
@@ -71,7 +73,7 @@ class transaksi_penjualancontroller extends Controller
         $data['bayar'] = str_replace(',', '', $request['bayar']);
         $data['kembali'] = str_replace(',', '', $request['kembali']);
 
-        $transaksiId = now() . transaksi_penjualan::all()->count();
+        $transaksiId = now()->format('dmy') . transaksi_penjualan::all()->count();
 
         transaksi_penjualan::where('no_transaction', $request['no_transaction'])
             ->update($data);
@@ -93,153 +95,153 @@ class transaksi_penjualancontroller extends Controller
         // return view('transaksi.penjualan.sales', compact('barangs'));
     }
 
-    public function storeDetail(DetailStoreRequest $request)
-    {
-        $input = $request->all();
+    // public function storeDetail(DetailStoreRequest $request)
+    // {
+    //     $input = $request->all();
 
-        $transaksiId = $input['no_transaction'];
-        $qty = $input['qty'];
+    //     $transaksiId = $input['no_transaction'];
+    //     $qty = $input['qty'];
 
-        transaksi_penjualan::firstOrCreate([
-            'no_transaction' => $transaksiId,
-        ]);
+    //     transaksi_penjualan::firstOrCreate([
+    //         'no_transaction' => $transaksiId,
+    //     ]);
 
-        $barangs = barang::where('no_barang', $input['no_barang'])->get();
-        foreach ($barangs as $barang) {
-            $barangId = $barang->id;
-            $barangName = $barang->name_barang;
-            $barangPrice = $barang->harga_jual;
-            $barangStock = $barang->stok;
-        }
+    //     $barangs = barang::where('no_barang', $input['no_barang'])->get();
+    //     foreach ($barangs as $barang) {
+    //         $barangId = $barang->id;
+    //         $barangName = $barang->name_barang;
+    //         $barangPrice = $barang->harga_jual;
+    //         $barangStock = $barang->stok;
+    //     }
 
-        if (isset($barangId)) {
-            return redirect()->back()->withErrors('Produk Tidak Ditemukan');
-        }
+    //     if (isset($barangId)) {
+    //         return redirect()->back()->withErrors('Produk Tidak Ditemukan');
+    //     }
 
-        $details = detail_penjualan::where([
-            ['transaksi_penjualans_id', '=', $transaksiId],
-            ['barangs_id', '=', $barangId]
-        ])->get();
+    //     $details = detail_penjualan::where([
+    //         ['transaksi_penjualans_id', '=', $transaksiId],
+    //         ['barangs_id', '=', $barangId]
+    //     ])->get();
 
-        $subTotal = $qty * $barangPrice;
-        $reduceStock = $barangStock - $qty;
+    //     $subTotal = $qty * $barangPrice;
+    //     $reduceStock = $barangStock - $qty;
 
-        $stockReduce = [
-            'stok' => $reduceStock
-        ];
+    //     $stockReduce = [
+    //         'stok' => $reduceStock
+    //     ];
 
-        $create = ([
-            'barangs_id' => $barangId,
-            'transaksi_penjualan_id' => $transaksiId,
-            'qty' => $qty,
-            'subTotal' => $subTotal,
-        ]);
+    //     $create = ([
+    //         'barangs_id' => $barangId,
+    //         'transaksi_penjualan_id' => $transaksiId,
+    //         'qty' => $qty,
+    //         'subTotal' => $subTotal
+    //     ]);
 
-        if ($barangStock == 0) {
-            return redirect()->back()->withErrors('Stok' . $barangName . 'Kosong');
-        }elseif ((int)$qty <= $barangStock) {
-            if ($details->isEmpty()) {
-                foreach ($details as $detail) {
-                    if ($detail->barangs_id == $barangId) {
-                        $update = [
-                            'qty' => $detail->qty + $qty,
-                            'subTotal' => $detail->subTotal + $subTotal,
-                        ];
-                        detail_penjualan::findOrFail($detail->id)->update($update);
-                        barang::findOrFail($barangId)->update($stockReduce);
-                    }else {
-                        detail_penjualan::create($create);
-                        barang::findOrFail($barangId)->update($stockReduce);
-                    }
+    //     if ($barangStock == 0) {
+    //         return redirect()->back()->withErrors('Stok' . $barangName . 'Kosong');
+    //     }elseif ((int)$qty <= $barangStock) {
+    //         if ($details->isEmpty()) {
+    //             foreach ($details as $detail) {
+    //                 if ($detail->barangs_id == $barangId) {
+    //                     $update = [
+    //                         'qty' => $detail->qty + $qty,
+    //                         'subTotal' => $detail->subTotal + $subTotal,
+    //                     ];
+    //                     detail_penjualan::findOrFail($detail->id)->update($update);
+    //                     barang::findOrFail($barangId)->update($stockReduce);
+    //                 }else {
+    //                     detail_penjualan::create($create);
+    //                     barang::findOrFail($barangId)->update($stockReduce);
+    //                 }
                     
-                }
-            }else {
-                detail_penjualan::create($create);
-                barang::findOrFail($barangId)->update($stockReduce);
-            }
-            return redirect()->route('transaksi.create', $transaksiId);
-        }else{
-            return redirect()->back();
-        }
+    //             }
+    //         }else {
+    //             detail_penjualan::create($create);
+    //             barang::findOrFail($barangId)->update($stockReduce);
+    //         }
+    //         return redirect()->route('transaksi.create', $transaksiId);
+    //     }else{
+    //         return redirect()->back();
+    //     }
 
        
 
         
 
-        // $sales = DB::table('transaksi_penjualans')->select(DB::raw('MAX(id) as num'));
-        // if ($sales->count() > 0) {
-        //     foreach ($sales->get() as $key) {
-        //         $no = ((int) $key->num + 1);
-        //     }
-        // }else {
-        //     $no = 1;
-        // }
+    //     // $sales = DB::table('transaksi_penjualans')->select(DB::raw('MAX(id) as num'));
+    //     // if ($sales->count() > 0) {
+    //     //     foreach ($sales->get() as $key) {
+    //     //         $no = ((int) $key->num + 1);
+    //     //     }
+    //     // }else {
+    //     //     $no = 1;
+    //     // }
 
         
 
         
-    }
+    // }
 
-    public function updateDetail(Request $request, $id)
-    {
-        $input = $request->all();
+    // public function updateDetail(Request $request, $id)
+    // {
+    //     $input = $request->all();
 
-        $id = $id;
+    //     $id = $id;
 
-        $transaksiId = $input['no_transaction'];
-        $qty = $input['qty'];
+    //     $transaksiId = $input['no_transaction'];
+    //     $qty = $input['qty'];
 
-        $valid = Validator::make($input, [
-            'no_transaction' => 'required',
-            'qty' => 'required'
-        ]);
+    //     $valid = Validator::make($input, [
+    //         'no_transaction' => 'required',
+    //         'qty' => 'required'
+    //     ]);
 
-        if ($valid->fails()) {
-            return redirect()->route('transaksi.create', $transaksiId)
-                ->withErrors($valid)
-                ->withInput();
-        }
+    //     if ($valid->fails()) {
+    //         return redirect()->route('transaksi.create', $transaksiId)
+    //             ->withErrors($valid)
+    //             ->withInput();
+    //     }
 
-        $details = detail_penjualan::find($id);
-        $stockSale = $details->qty;
-        $barangId = $details->barangs_id;
+    //     $details = detail_penjualan::find($id);
+    //     $stockSale = $details->qty;
+    //     $barangId = $details->barangs_id;
 
-        $barangs = barang::findOrFail($barangId);
-        $barangPrice = $barangs->harga_jual;
-        $barangStock = $barangs->stok;
+    //     $barangs = barang::findOrFail($barangId);
+    //     $barangPrice = $barangs->harga_jual;
+    //     $barangStock = $barangs->stok;
 
-        $firstStock = $barangStock + $stockSale;
-        $stockReduce = $firstStock - $qty;
-        $total = $qty * $barangPrice;
+    //     $firstStock = $barangStock + $stockSale;
+    //     $stockReduce = $firstStock - $qty;
+    //     $total = $qty * $barangPrice;
 
-        if ((int)$qty < $firstStock) {
-            detail_penjualan::findOrFail($id)->update([
-                'qty' => $qty,
-                'subTotal' => $total
-            ]);
-            barang::findOrFail($barangId)->update([
-                'stok' => $stockReduce
-            ]);
-            return redirect()->route('transaksi.create');
-        }else {
-            return redirect()->back()->withErrors('Stok tidak mencukupi');
-        }
-    }
+    //     if ((int)$qty < $firstStock) {
+    //         detail_penjualan::findOrFail($id)->update([
+    //             'qty' => $qty,
+    //             'subTotal' => $total
+    //         ]);
+    //         barang::findOrFail($barangId)->update([
+    //             'stok' => $stockReduce
+    //         ]);
+    //         return redirect()->route('transaksi.create');
+    //     }else {
+    //         return redirect()->back()->withErrors('Stok tidak mencukupi');
+    //     }
+    // }
 
-    public function destroy($id)
-    {
-        $details = detail_penjualan::findOrFail($id);
-        $stockSale = $details->qty;
-        $barangId = $details->barangs_id;
+    // public function destroy($id)
+    // {
+    //     $details = detail_penjualan::findOrFail($id);
+    //     $stockSale = $details->qty;
+    //     $barangId = $details->barangs_id;
 
-        $barangStock = barang::findOrFail($barangId)->stok;
-        $firstStock = $barangStock + $stockSale;
+    //     $barangStock = barang::findOrFail($barangId)->stok;
+    //     $firstStock = $barangStock + $stockSale;
 
-        barang::findOrFail($barangId)->update([
-            'stok' => $firstStock
-        ]);
+    //     barang::findOrFail($barangId)->update([
+    //         'stok' => $firstStock
+    //     ]);
 
-        detail_penjualan::findOrFail($id)->delete();
-        return redirect()->back();
-    }
+    //     detail_penjualan::findOrFail($id)->delete();
+    //     return redirect()->back();
+    // }
 }
