+++ app/Http/Requests/DetailStoreRequest.php
@@ -13,7 +13,7 @@ class DetailStoreRequest extends FormRequest
      */
     public function authorize()
     {
-        return false;
+        return true;
     }
 
     /**
